Moved localglobalbottombarheight to Locals.kt
