+++ composeApp/src/commonMain/kotlin/paige/navic/util/ui/PaddingUtils.kt
@@ -1,13 +1,9 @@
 package paige.navic.util.ui
 
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.calculateEndPadding
-import androidx.compose.foundation.layout.calculateStartPadding
 import androidx.compose.foundation.layout.minus
 import androidx.compose.foundation.layout.plus
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalLayoutDirection
-import paige.navic.LocalGlobalBottomBarHeight
 
 fun PaddingValues.withoutTop() = this.minus(PaddingValues(top = this.calculateTopPadding()))
